% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{US_Trade_Deficit}
\alias{US_Trade_Deficit}
\title{US Trade Deficit Data: 1987-1988.}
\format{
A data frame with 24 rows and 2 variables:
\describe{
  \item{date}{observation month}
  \item{deficit_billions}{US trade deficit in billions of dollars}
  ...
}
}
\usage{
US_Trade_Deficit
}
\description{
A replication of the US Trade Deficit data used in \href{https://variation.com/wp-content/uploads/change-point-analyzer/change-point-analysis-a-powerful-new-tool-for-detecting-changes.pdf}{Taylor's manuscript}.
}
\keyword{datasets}
